third_party/blink/web_tests/virtual/disable-single-axis-scroll-containers/external/wpt/css/css-position/sticky/position-sticky-single-axis-nested-expected.txt

This is a testharness.js-based test.
[FAIL] Nested sticky elements accumulate offsets correctly on independent axes
  assert_equals: Outer sticky top expected 10 but got -50
Harness: the test ran to completion.
